io_uring: support user sqe ext flags
sqe->flags is u8, and now we have used 7 bits, so take the last one for
extending purpose.

If bit8(IOSQE_HAS_EXT_FLAGS_BIT) is defined, it means this sqe carries
ext flags from the last byte(.ext_flags), or bit23~bit16 of sqe->uring_cmd_flags
for IORING_OP_URING_CMD.

io_slot_flags() return value is converted to `ULL` because the affected
bits is beyond 32bit now.

+++ io_uring/filetable.h
@@ -43,7 +43,7 @@ io_fixed_file_slot(struct io_file_table *table, unsigned i)
 #define FFS_ISREG		0x2UL
 #define FFS_MASK		~(FFS_NOWAIT|FFS_ISREG)
 
-static inline unsigned int io_slot_flags(struct io_fixed_file *slot)
+static inline unsigned long io_slot_flags(struct io_fixed_file *slot)
 {
 	return (slot->file_ptr & ~FFS_MASK) << REQ_F_SUPPORT_NOWAIT_BIT;
 }
